Add Bottom Sheet Report Fragment to display disaster data
This commit adds a new Bottom Sheet Dialog Fragment named "ReportFragment" to display disaster data. The fragment is designed to be used as a bottom sheet, providing a smooth user experience.

The ReportFragment utilizes the ReportViewModel to handle data logic and fetch disaster reports using the UseCase from the Core module. The ViewModel communicates with the UseCase to retrieve the disaster data from the data source.

The RecyclerView in the ReportFragment displays the fetched disaster reports, and a custom ReportAdapter is used to efficiently handle the list of data.

+++ coreapp/src/main/java/com/hann/disasterguard/coreapp/domain/model/GeometryReport.kt
@@ -3,7 +3,7 @@ package com.hann.disasterguard.coreapp.domain.model
 import com.hann.disasterguard.coreapp.data.remote.response.report.Properties
 
 data class GeometryReport(
-    val coordinates: List<Int>,
+    val coordinates: List<Double>,
     val properties: Properties,
     val type: String
 )
